+++ .github/workflows/code-quality.yaml
@@ -22,7 +22,6 @@ jobs:
     strategy:
       matrix:
         python_version:
-        - "3.9"
         - "3.10"
         - "3.11"
         - "3.12"
