+++ .github/workflows/pr-docker.yaml
@@ -23,7 +23,7 @@ jobs:
     steps:
     - uses: actions/setup-python@v4
       with:
-        python-version: 3.9
+        python-version: 3.10
     - uses: actions/checkout@v3
     - id: set-matrix
       run: |
