+++ .github/workflows/release.yaml
@@ -12,7 +12,6 @@ jobs:
     strategy:
       matrix:
         python_version:
-        - "3.9"
         - "3.10"
         - "3.11"
         - "3.12"
@@ -44,7 +43,7 @@ jobs:
     - name: Set up Python
       uses: actions/setup-python@v3
       with:
-        python-version: "3.9"
+        python-version: "3.10"
 
     - name: Build source and wheel distributions
       run: |
@@ -86,7 +85,7 @@ jobs:
     - name: Set up Python
       uses: actions/setup-python@v3
       with:
-        python-version: "3.9"
+        python-version: "3.10"
 
     - name: Build source and wheel distributions
       run: |
