+++ .github/workflows/smoketest.yaml
@@ -23,7 +23,7 @@ jobs:
     strategy:
       matrix:
         python_version:
-        - "3.9"
+        - "3.10"
         - "3.10"
         - "3.11"
         - "3.12"
