+++ tests/fixtures/fixtures.py
@@ -485,3 +485,29 @@ def tiny_mpt_tokenizer(_session_tiny_mpt_tokenizer):
 @pytest.fixture
 def tiny_mpt_model(_session_tiny_mpt_model):
     return copy.deepcopy(_session_tiny_mpt_model)
+
+
+@pytest.fixture
+def clean_mlflow_runs():
+    """Clean up MLflow runs before and after tests.
+
+    This fixture ensures no MLflow runs persist between tests,
+    which prevents "Run already active" errors.
+    """
+    try:
+        import mlflow
+        try:
+            while mlflow.active_run():
+                mlflow.end_run()
+        except Exception:
+            pass
+
+        yield
+
+        try:
+            while mlflow.active_run():
+                mlflow.end_run()
+        except Exception:
+            pass
+    except ImportError:
+        yield
